requirements.txt
python-dotenv~=1.0.1
envparse~=0.2.0
whisper==1.1.10
python-ffmpeg
openai-whisper